refactor: move handlers to src folder
