+++ product-service/src/handlers/getProductById.ts
