+++ product-service/src/handlers/getProductsList.ts
