+++ product-service/src/handlers/handler.ts
